Abstract bpf_exit
Separate the functionality of bpf_exit into a header file.
